% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fGenerateQ}
\alias{fGenerateQ}
\title{Draw a Random Orthonormal Matrix (RZW 2010)}
\usage{
fGenerateQ(N)
}
\arguments{
\item{N}{Integer. Dimension of the square matrix.}
}
\value{
An N x N orthonormal matrix.
}
\description{
Generates an N x N orthonormal matrix Q (QQ' = Q'Q = I) by applying QR
decomposition to a matrix of i.i.d. standard normals and normalising the
diagonal of R to be positive, yielding a unique decomposition.
}
\details{
Draw an N x N matrix M of independent standard normals, compute the
QR decomposition M = QR, then set
Q[,i] = -Q[,i] whenever R[i,i] < 0.  The resulting Q is uniformly
distributed on the Stiefel manifold (Haar measure), as required for
sign-restriction identification.
}
\examples{
\dontrun{
Q <- fGenerateQ(3)
round(t(Q) \%*\% Q, 10)  # should be identity
}

}
\references{
Rubio-Ramirez, J. F., Waggoner, D. F., & Zha, T. (2010). Structural vector
autoregressions: Theory of identification and algorithms for inference.
\emph{Review of Economic Studies}, 77(2), 665--696.
}
