% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fPolyConvolve}
\alias{fPolyConvolve}
\title{Polynomial Convolution of Two Matrix Lag Polynomials}
\usage{
fPolyConvolve(A, B, nlags)
}
\arguments{
\item{A}{3-D array of dimension \eqn{d_1 \times d_2 \times A_3}
(e.g. Wold IRF coefficients from \code{fwoldIRF}).}

\item{B}{3-D array of dimension \eqn{d_2 \times K \times B_3}
(e.g. \eqn{\Psi(L)} coefficients).}

\item{nlags}{Integer. Number of output slices to return
(must be \eqn{\leq A_3 + B_3 - 1}).}
}
\value{
A 3-D array of dimension \eqn{d_1 \times K \times} \code{nlags}.
}
\description{
Computes \eqn{C(L) = A(L) \cdot B(L)} truncated to \code{nlags} terms.
Both polynomials are stored as 3-D arrays where the third dimension indexes
lag order (slice 1 = lag 0, slice 2 = lag 1, etc.).
}
\seealso{
\code{\link{fwoldIRF}}, \code{\link{fcholeskyIRF}}
}
