% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyMacro-package.R
\docType{package}
\name{tidyMacro-package}
\alias{tidyMacro}
\alias{tidyMacro-package}
\title{tidyMacro: Fast Vector Autoregression Models via C++}
\description{
Implements Vector Autoregression (VAR) models using fast and efficient C++ programs via Rcpp and RcppArmadillo. Provides tools for estimation, forecasting, and impulse response analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/muhsinciftci/tidyMacro}
  \item Report bugs at \url{https://github.com/muhsinciftci/tidyMacro/issues}
}

}
\author{
\strong{Maintainer}: Muhsin Ciftci \email{farmer.muhsin@gmail.com}

}
